style: minor improvements

+++ src/socket/tcp.rs
@@ -181,7 +181,7 @@ impl TcpSocket {
                 self.send_flags.as_i32(),
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
             self.buffer.shift_left_buffer(result as usize);
