+++ src/socket/udp.rs
@@ -185,7 +185,7 @@ impl UdpSocket {
                 self.recv_flags.as_i32(),
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
             Ok(result as usize)
@@ -220,7 +220,7 @@ impl UdpSocket {
                 &mut Self::socket_len(),
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
             Ok(result as usize)
@@ -268,7 +268,7 @@ impl UdpSocket {
                 socklen,
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
             self.buffer.shift_left_buffer(result as usize);
@@ -321,7 +321,7 @@ impl UdpSocket {
                 self.send_flags.as_i32(),
             )
         };
-        if (result as i32) < 0 {
+        if result < 0 {
             Err(SocketError::Errno(unsafe { sys::sceNetInetGetErrno() }))
         } else {
             self.buffer.shift_left_buffer(result as usize);
